croupier: Modify Game UI to get nearer to true Hearts UI
The Card Collection is now drawn more nicely (Stack of Positioned Card).
I use Flex with JustifyFlexContent.spaceAround and spaceBetween.

The UI is kinda like this:
https://drive.google.com/a/google.com/file/d/0B8atBVyuAi_kbDZlekhSeGdNaFk/view?usp=sharing

With a little bit of extra debug info.

Since a lot of the components now need to know their size, I have started
to track widths and heights, somewhat messily.

Change-Id: I02b6a1376d35cb776bd36c6b77a7f8e38ef19931