croupier: Modify commands and add syncbase log schema
Commands now identify a phase and a the level of simultaneity.

Depending on whether or not something is simultaneous, the LogWriter (or mock)
will act differently.

The mock doesn't need to do much extra work since it "knows" it's on a single
device. However, it's implementation closely mirrors that of the Syncbase
Schema. The mock has been well-tested, as has the Syncbase schema.

The procedure was basically to play a complete round in each case and to try
out "bad moves" and check that there were no crashes or deadlocks.

Next steps:
- Create a Multiplayer Hearts Game that forks off the current one.
  This one may require using 4 Android or 4 Desktop instances.
  Both are problematic, so this will require some thought.
- While blocked on the former, I can also work on cleaning up the code
  on the implementation and UI sides.

Change-Id: I23cd49e0552287e03ad2eb787f15cb5d92e88dc8