croupier: Introduce Settings Manager
The Settings manager describes features of the Croupier player.
This information is stored in Syncbase and in the future, will be synced
across players of the same game.

To assist with that, the LogWriter was partially abstracted out to a
syncbase CroupierClient. A mock is also created for tests.

The CroupierSettings include a user id, avatar, color, and name.
Only the last 3 can be modified by the user.

Due to some issues with Sky, only the user's name can be changed
(in a rudimentary UI). The user's other settings can be viewed
(excluding the user id).

Change-Id: If12383f9f255a95a3648c079c5ef1a66204055b7