croupier: Croupier with 2+ devices
The Makefile setup has been upgraded (accompanied by README.md) such that
we can run Croupier on multiple Android devices at the same time.

To further prove this, the devices join a syncgroup together and
share each other's settings. These settings are shown (in an ugly
manner) as CroupierProfileComponents.

There are some caveats though:
- must change Mojo to expose some ports as environment variables
- must start the first device (syncgroup) early enough
- that same first device's app will likely crash after restarting the
  app, so it must be fully uninstalled.
- must connect the devices in a specific order and use ANDROID=1 (or 2, 3, ...)
  to reference them. (We parse adb logcat in order to get the device id.)

These Syncgroups never forget devices (at this point in time), so this is
not necessarily equivalent to Discovery. However, it is close enough that
we can continue onto the Invite Players and Arrange Players screens.

Change-Id: I8f940930fddcbb81fe7df408c3bacc84cd61eb54