croupier: Add Card Transitions Layer
(UI) Cards now have the option of being invisible. If they are invisible, it
is up to the Game to capture the Cards that are rendered and compute their
positions.

This asynchronously collected data will be used to render a new Card layer
on top of the rest of the game widgets. The previously invisible Cards are
rendered here as ZCards.

ZCards mirror Cards except that they keep track of a queue of positions to
animate to. This allows them to move across the screen as actions take place.

ZCards and their wrapping Positioned are keyed widgets. Since they have keys,
they are not allowed to be duplicated. Therefore, extra work will need to be
done to have the same underlying logic card animate in two ways.

ZCards also respect the z-index given to them. Right now, this value is given
naively, but there can be superior strategies in the future.

Note: Cards that are drag-and-dropped still animate.
This can probably be avoided in the future by ensuring the callback is aware
of the dropped card(s).

Note: Cards don't appear in the right places sometimes since their 'keyed'ness
makes them "remember" their old offset. This causes it to apply "twice" or
more in some scenarios. https://github.com/flutter/engine/issues/1939

Change-Id: Ib88a1c697c63b06ef1f0cda1d8ffdbed61583020