croupier: Upgrade to the real v23discovery
Croupier has been a little behind on discovery, being the first ones
to use it since before discovery was published. Since it is now
v23discovery, we no longer need to rely on local .mojo paths.

pubspec.lock was also added in this CL, since Dart applications
are meant to check those in.

Change-Id: Ie3e315fb3bf1a4729a661e50bbc4a8eb54ae88da