croupier: Sort Watch Updates to LogWriter
Only the LogWriter needs this since it cares heavily about the order
in which things are updated.

Tested by playing Solitaire with another peer and making many quick
moves. Originally, this would eventually lead to a failure (due to
not sorting the watch updates). But now, it plays just fine.

Closes https://github.com/vanadium/issues/issues/844

Change-Id: If7660bbfbe7d7992f90641afcf497d55bdbbff8e