croupier: Add Shortcut creation logic to Makefile
This adds a way to create shortcuts to the Makefile.
While this does belong in mojo.shared, this is here because of the
recent refactoring done there. The next CL will MultiPart the relevant
repositories.

mojo.shared's macro will need to take 5 variables:
* Target device ID Flag (optional)
* Shortcut Name (required)
* Shortcut Mojo Url (required)
* Shortcut Icon Url (optional)
* Shell command path (required)

https://github.com/vanadium/issues/issues/942

The Shortcut also requires some assets to be loaded onto a public and
secure location, like storage.googleapis.com. For now, I have manually
put a few of our binaries up there, though in the future we will also
want to version them.

https://github.com/vanadium/issues/issues/834

The related files are:
* Makefile
* shortcut_template
* .gitignore

=========================

The other files are actually part of a related change.

This other half modifies the Hearts and Croupier Game implementations.
A new signal called startGameSignal is called at the start of a game.
This allows Hearts to send a Ready command when the button is pressed.
This allows Deal to occur independently of the other players, since the
owner is known, and we can let them deal.
The tests were updated to correspond with this change.

Change-Id: I1acf494c45d7d3d9dda3acba815a3fe63a13e966