croupier: Drag And Drop Arrange Players
Now allows players to drag and drop themselves (and others) into the
appropriate slots.

You can also move people that are in the way.

This isn't very good when more than 1 person ends up in the same slot;
however, data consistency is maintained in Syncbase.

Note: We always had this race condition, so this isn't a new problem.
This CL alleviates the problem since you can just reposition yourself
when such a thing happens.

Change-Id: I4937c3f6e6f9b8e7a55b567dd02928fcfdf5f12a